AST code uses PropertyFetching techniques  (#1742)
* Made the AST code use PropertyFetching techniques to get pojo values and turn them into AST elements

* Fixed Map support by moving it down into the common code
