+++ src/main/java/graphql/schema/PropertyDataFetcher.java
@@ -2,26 +2,10 @@ package graphql.schema;
 
 
 import graphql.Assert;
-import graphql.GraphQLException;
 import graphql.PublicApi;
 import graphql.TrivialDataFetcher;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
-
-import static graphql.Scalars.GraphQLBoolean;
-import static graphql.schema.GraphQLTypeUtil.isNonNull;
-import static graphql.schema.GraphQLTypeUtil.unwrapOne;
 
 /**
  * This is the default data fetcher used in graphql-java.  It will examine
@@ -51,13 +35,6 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
 
     private final String propertyName;
     private final Function<Object, Object> function;
-
-    private static final AtomicBoolean USE_SET_ACCESSIBLE = new AtomicBoolean(true);
-    private static final AtomicBoolean USE_NEGATIVE_CACHE = new AtomicBoolean(true);
-    private static final ConcurrentMap<String, Method> METHOD_CACHE = new ConcurrentHashMap<>();
-    private static final ConcurrentMap<String, Field> FIELD_CACHE = new ConcurrentHashMap<>();
-    private static final ConcurrentMap<String, String> NEGATIVE_CACHE = new ConcurrentHashMap<>();
-
 
     /**
      * This constructor will use the property name and examine the {@link DataFetchingEnvironment#getSource()}
@@ -92,7 +69,6 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
      *
      * @param propertyName the name of the property to retrieve
      * @param <T>          the type of result
-     *
      * @return a new PropertyDataFetcher using the provided function as its source of values
      */
     public static <T> PropertyDataFetcher<T> fetching(String propertyName) {
@@ -116,7 +92,6 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
      * @param function the function to use to obtain a value from the source object
      * @param <O>      the type of the source object
      * @param <T>      the type of result
-     *
      * @return a new PropertyDataFetcher using the provided function as its source of values
      */
     public static <T, O> PropertyDataFetcher<T> fetching(Function<O, T> function) {
@@ -142,87 +117,7 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
             return (T) function.apply(source);
         }
 
-        if (source instanceof Map) {
-            return (T) ((Map<?, ?>) source).get(propertyName);
-        }
-        return (T) getPropertyViaGetter(source, environment.getFieldType(), environment);
-    }
-
-    private Object getPropertyViaGetter(Object object, GraphQLOutputType outputType, DataFetchingEnvironment environment) {
-        String key = mkKey(object);
-        if (isNegativelyCached(key)) {
-            return null;
-        }
-        try {
-            return getPropertyViaGetterMethod(object, outputType, this::findPubliclyAccessibleMethod, environment);
-        } catch (NoSuchMethodException ignored) {
-            try {
-                return getPropertyViaGetterMethod(object, outputType, this::findViaSetAccessible, environment);
-            } catch (NoSuchMethodException ignored2) {
-                try {
-                    return getPropertyViaFieldAccess(object);
-                } catch (FastNoSuchMethodException e) {
-                    // we have nothing to ask for and we have exhausted our lookup strategies
-                    putInNegativeCache(key);
-                    return null;
-                }
-            }
-        }
-    }
-
-    private boolean isNegativelyCached(String key) {
-        if (USE_NEGATIVE_CACHE.get()) {
-            return NEGATIVE_CACHE.containsKey(key);
-        }
-        return false;
-    }
-
-    private void putInNegativeCache(String key) {
-        if (USE_NEGATIVE_CACHE.get()) {
-            NEGATIVE_CACHE.put(key, key);
-        }
-    }
-
-    @FunctionalInterface
-    private interface MethodFinder {
-        Method apply(Class aClass, String s) throws NoSuchMethodException;
-    }
-
-    private Object getPropertyViaGetterMethod(Object object, GraphQLOutputType outputType, MethodFinder methodFinder, DataFetchingEnvironment environment) throws NoSuchMethodException {
-        if (isBooleanProperty(outputType)) {
-            try {
-                return getPropertyViaGetterUsingPrefix(object, "is", methodFinder, environment);
-            } catch (NoSuchMethodException e) {
-                return getPropertyViaGetterUsingPrefix(object, "get", methodFinder, environment);
-            }
-        } else {
-            return getPropertyViaGetterUsingPrefix(object, "get", methodFinder, environment);
-        }
-    }
-
-    private Object getPropertyViaGetterUsingPrefix(Object object, String prefix, MethodFinder methodFinder, DataFetchingEnvironment environment) throws NoSuchMethodException {
-        String getterName = prefix + propertyName.substring(0, 1).toUpperCase() + propertyName.substring(1);
-        try {
-            Method method = methodFinder.apply(object.getClass(), getterName);
-            if (takesDataFetcherEnvironmentAsOnlyArgument(method)) {
-                return method.invoke(object, environment);
-            } else {
-                return method.invoke(object);
-            }
-        } catch (IllegalAccessException | InvocationTargetException e) {
-            throw new GraphQLException(e);
-        }
-    }
-
-    @SuppressWarnings("SimplifiableIfStatement")
-    private boolean isBooleanProperty(GraphQLOutputType outputType) {
-        if (outputType == GraphQLBoolean) {
-            return true;
-        }
-        if (isNonNull(outputType)) {
-            return unwrapOne(outputType) == GraphQLBoolean;
-        }
-        return false;
+        return (T) PropertyDataFetcherHelper.getPropertyValue(propertyName, source, environment.getFieldType(), environment);
     }
 
     /**
@@ -235,9 +130,7 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
      */
     @SuppressWarnings("unused")
     public static void clearReflectionCache() {
-        METHOD_CACHE.clear();
-        FIELD_CACHE.clear();
-        NEGATIVE_CACHE.clear();
+        PropertyDataFetcherHelper.clearReflectionCache();
     }
 
     /**
@@ -245,157 +138,19 @@ public class PropertyDataFetcher<T> implements DataFetcher<T>, TrivialDataFetche
      * values.  By default it PropertyDataFetcher WILL use setAccessible.
      *
      * @param flag whether to use setAccessible
-     *
      * @return the previous value of the flag
      */
     public static boolean setUseSetAccessible(boolean flag) {
-        return USE_SET_ACCESSIBLE.getAndSet(flag);
+        return PropertyDataFetcherHelper.setUseSetAccessible(flag);
     }
 
     /**
      * This can be used to control whether PropertyDataFetcher will cache negative lookups for a property for performance reasons.  By default it PropertyDataFetcher WILL cache misses.
      *
      * @param flag whether to cache misses
-     *
      * @return the previous value of the flag
      */
     public static boolean setUseNegativeCache(boolean flag) {
-        return USE_NEGATIVE_CACHE.getAndSet(flag);
-    }
-
-    private String mkKey(Object object) {
-        return mkKey(object.getClass(), propertyName);
-    }
-
-    private String mkKey(Class clazz, String propertyName) {
-        return clazz.getCanonicalName() + "__" + propertyName;
-    }
-
-    // by not filling out the stack trace, we gain speed when using the exception as flow control
-    @SuppressWarnings("serial")
-    private static class FastNoSuchMethodException extends NoSuchMethodException {
-        public FastNoSuchMethodException(String methodName) {
-            super(methodName);
-        }
-
-        @Override
-        public synchronized Throwable fillInStackTrace() {
-            return this;
-        }
-    }
-
-    /**
-     * Invoking public methods on package-protected classes via reflection
-     * causes exceptions. This method searches a class's hierarchy for
-     * public visibility parent classes with the desired getter. This
-     * particular case is required to support AutoValue style data classes,
-     * which have abstract public interfaces implemented by package-protected
-     * (generated) subclasses.
-     */
-    @SuppressWarnings("unchecked")
-    private Method findPubliclyAccessibleMethod(Class root, String methodName) throws NoSuchMethodException {
-        Class currentClass = root;
-        while (currentClass != null) {
-            String key = mkKey(currentClass, propertyName);
-            Method method = METHOD_CACHE.get(key);
-            if (method != null) {
-                return method;
-            }
-            if (Modifier.isPublic(currentClass.getModifiers())) {
-                //
-                // try a getter that takes DataFetchingEnvironment first
-                try {
-                    method = currentClass.getMethod(methodName, DataFetchingEnvironment.class);
-                    if (Modifier.isPublic(method.getModifiers())) {
-                        METHOD_CACHE.putIfAbsent(key, method);
-                        return method;
-                    }
-                } catch (NoSuchMethodException e) {
-                    // ok try the next approach
-                }
-                method = currentClass.getMethod(methodName);
-                if (Modifier.isPublic(method.getModifiers())) {
-                    METHOD_CACHE.putIfAbsent(key, method);
-                    return method;
-                }
-            }
-            currentClass = currentClass.getSuperclass();
-        }
-        return root.getMethod(methodName);
-    }
-
-    private Method findViaSetAccessible(Class aClass, String methodName) throws NoSuchMethodException {
-        if (!USE_SET_ACCESSIBLE.get()) {
-            throw new FastNoSuchMethodException(methodName);
-        }
-        Class currentClass = aClass;
-        while (currentClass != null) {
-            String key = mkKey(currentClass, propertyName);
-            Method method = METHOD_CACHE.get(key);
-            if (method != null) {
-                return method;
-            }
-
-            Method[] declaredMethods = currentClass.getDeclaredMethods();
-            Optional<Method> m = Arrays.stream(declaredMethods)
-                    .filter(mth -> methodName.equals(mth.getName()))
-                    .filter(mth -> hasZeroArgs(mth) || takesDataFetcherEnvironmentAsOnlyArgument(mth))
-                    .min(mostMethodArgsFirst());
-            if (m.isPresent()) {
-                try {
-                    // few JVMs actually enforce this but it might happen
-                    method = m.get();
-                    method.setAccessible(true);
-                    METHOD_CACHE.putIfAbsent(key, method);
-                    return method;
-                } catch (SecurityException ignored) {
-                }
-            }
-            currentClass = currentClass.getSuperclass();
-        }
-        throw new FastNoSuchMethodException(methodName);
-    }
-
-    private boolean hasZeroArgs(Method mth) {
-        return mth.getParameterCount() == 0;
-    }
-
-    private boolean takesDataFetcherEnvironmentAsOnlyArgument(Method mth) {
-        return mth.getParameterCount() == 1 &&
-                mth.getParameterTypes()[0].equals(DataFetchingEnvironment.class);
-    }
-
-    private Comparator<? super Method> mostMethodArgsFirst() {
-        return Comparator.comparingInt(Method::getParameterCount).reversed();
-    }
-
-    private Object getPropertyViaFieldAccess(Object object) throws FastNoSuchMethodException {
-        Class<?> aClass = object.getClass();
-        String key = mkKey(aClass, propertyName);
-        try {
-            Field field = FIELD_CACHE.get(key);
-            if (field == null) {
-                field = aClass.getField(propertyName);
-                FIELD_CACHE.putIfAbsent(key, field);
-            }
-            return field.get(object);
-        } catch (NoSuchFieldException e) {
-            if (!USE_SET_ACCESSIBLE.get()) {
-                throw new FastNoSuchMethodException(key);
-            }
-            // if not public fields then try via setAccessible
-            try {
-                Field field = aClass.getDeclaredField(propertyName);
-                field.setAccessible(true);
-                FIELD_CACHE.putIfAbsent(key, field);
-                return field.get(object);
-            } catch (SecurityException | NoSuchFieldException ignored2) {
-                throw new FastNoSuchMethodException(key);
-            } catch (IllegalAccessException e1) {
-                throw new GraphQLException(e);
-            }
-        } catch (IllegalAccessException e) {
-            throw new GraphQLException(e);
-        }
+        return PropertyDataFetcherHelper.setUseNegativeCache(flag);
     }
 }
